pyproject.toml
[project]
name = "rag-qa-system"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chromadb>=1.0.20",
    "faiss-cpu>=1.12.0",
    "feedparser>=6.0.12",
    "ipykernel>=6.30.1",
    "langchain>=0.3.27",
    "langchain-community>=0.3.29",
    "langchain-core>=0.3.76",
    "langchain-groq>=0.3.8",
    "pymupdf>=1.26.4",
    "pypdf>=6.0.0",
    "python-dotenv>=1.1.1",
    "sentence-transformers>=5.1.0",
    "streamlit>=1.49.1",
    "streamlit-tags>=1.2.8",
    "annoy>=1.17.3",
    "scann>=1.3.2",
    "pandas>=2.0.0",
    "matplotlib>=3.7.0",
    "psutil>=5.9.0",
    "numpy>=1.24.0",
]
